% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{blank.remover}
\alias{blank.remover}
\title{Remove blanks in the end of characters}
\usage{
blank.remover(vec.char)
}
\arguments{
\item{vec.char}{A vector of characters for testing}
}
\value{
A vector of characters
}
\description{
Remove blanks in the end of characters
}
\note{
This function detects blanks in the end of characters and
             delete them
}
\examples{
vec.char = c("QWERTYUIOP   ",
             "ASDFGHJKL:  ",
             "ZXCVBNM<>? ")
blank.remover(vec.char)
# Blank found!
# Blank found!
# Blank found!
# [1] "QWERTYUIOP" "ASDFGHJKL:" "ZXCVBNM<>?"
}

