% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clc.functions}
\alias{clc.functions}
\title{clean all (customed) functions in present environment}
\usage{
clc.functions()
}
\note{
This function looks for all functions defined in current 
             session and cleaning them. It may be useful when a 
             environment image is loaded and functions may be masked.
}
