% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_genomic_map.R
\name{create_genomic_map}
\alias{create_genomic_map}
\title{Create a geno.map object}
\usage{
create_genomic_map(SNPID, chromosome, position, autosome, allosome = NULL,
  mitochondria = NULL)
}
\arguments{
\item{SNPID}{a character vector of snp ID}

\item{chromosome}{a character vector of chromosome ID}

\item{position}{a integer vector of physical position}

\item{autosome}{a character vector of autosomes}

\item{allosome}{a character vector of allosomes, default: excluded}

\item{mitochondria}{a character vector of mitochondria, default: excluded}
}
\value{
a geno.map object
}
\description{
DESCRIPTION
}
\note{
A "factory unction" for geno.map object
}
