% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genomic_map.R
\name{CreateWindowMap}
\alias{CreateWindowMap}
\title{Create a windows map}
\usage{
CreateWindowMap(SID, chromosome, position, allosome = NULL, distance = 10^6)
}
\arguments{
\item{SID}{A vector of SNP IDs (character)}

\item{chromosome}{A vector of chromosome (character) where the SNP
is located. Support [1-9]{1,2}|[A-Z] ie. 0-99 or
a single upper letter. 0 for unmapped SNPs}

\item{position}{A vector of position (integer) where the SNP
is located.}

\item{allosome}{A vector of the names of allosomes. By default all
chromosomes natated with upper letters will be
treated as allosomes}

\item{distance}{A integer of window size}
}
\value{
A geno.map.window object
}
\description{
Create a windows map
}
\note{
This function create a windows map for "windows result" based
             original map.
}
\examples{
#' SID = paste("SNP",1:20,sep="_")
chromosome = c(rep("0",3),rep("1",5),rep("2",3),rep("X",6),rep("Y",3)))
position = sample(1:100,20)

data.frame(SID,chromosome,position)
#       SID chromosome position
# 1   SNP_1          0       31
# 2   SNP_2          0       71
# 3   SNP_3          0       45
# 4   SNP_4          1       60
# 5   SNP_5          1       34
# 6   SNP_6          1       41
# 7   SNP_7          1       19
# 8   SNP_8          1       23
# 9   SNP_9          2       63
# 10 SNP_10          2       43
# 11 SNP_11          2        2
# 12 SNP_12          X       79
# 13 SNP_13          X        6
# 14 SNP_14          X       95
# 15 SNP_15          X       39
# 16 SNP_16          X       97
# 17 SNP_17          X       15
# 18 SNP_18          Y       32
# 19 SNP_19          Y       56
# 20 SNP_20          Y       88

CreateWindowMap(SID, chromosome, position, distance = 30)
# $Chromosomes
# chr.id chr.is.allosome chr.length chr.n.window
# 0      0           FALSE         71            3
# 1      1           FALSE         60            2
# 2      2           FALSE         63            3
# X      X            TRUE         97            4
# Y      Y            TRUE         88            3
#
# $Windows
# $Windows$`0`
# ID chromosome position
# 1 0_1          0        0
# 2 0_2          0        1
# 3 0_3          0        2
#
# $Windows$`1`
# ID chromosome position
# 1 1_1          1        0
# 2 1_2          1        1
#
# $Windows$`2`
# ID chromosome position
# 1 2_1          2        0
# 2 2_2          2        1
# 3 2_3          2        2
#
# $Windows$X
# ID chromosome position
# 1 X_1          X        0
# 2 X_2          X        1
# 3 X_3          X        2
# 4 X_4          X        3
#
# $Windows$Y
# ID chromosome position
# 1 Y_1          Y        0
# 2 Y_2          Y        1
# 3 Y_3          Y        2
#
#
# attr(,"class")
# [1] "geno.map.window"
}

