% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{DrawAFrame}
\alias{DrawAFrame}
\title{Position calculation and frame plotting}
\usage{
DrawAFrame(map.sorted, gap, ylim, axes = FALSE, xlab = "", ylab = "",
  title = "")
}
\arguments{
\item{map.sorted}{A sorted map}

\item{gap}{The gap between chromosomes in the plot}

\item{ylim}{The y-axis range}

\item{axes}{Logical, draw the axes or not}

\item{xlab,ylab}{Parameters passed to plot.default()}

\item{title}{Character, the title of the plot}
}
\value{
A data frame of chromosome information
}
\description{
xlab = "", ylab = "", title = "")
}
\note{
This is a private function that calculate the genome positions
             and draw a basic frame for manhattan plots
}
\examples{
SID = paste("SNP",1:20,sep="_")
chromosome = c(rep("0",3),rep("1",5),rep("2",3),rep("X",6),rep("Y",3))
position = c(31, 71, 45, 60, 34, 41, 19, 23, 63, 43,
              2, 79,  6 ,95, 39, 97, 15, 32, 56, 88)

(map.sorted = SortMap(SID, chromosome, position))
# $Chromosomes
#   chr.id chr.is.allosome chr.length
# 0      0           FALSE         71
# 1      1           FALSE         60
# 2      2           FALSE         63
# X      X            TRUE         97
# Y      Y            TRUE         88
#
# $SNPs
# $SNPs$`0`
# ID chromosome position
# 1 SNP_1          0       31
# 3 SNP_3          0       45
# 2 SNP_2          0       71
#
# $SNPs$`1`
# ID chromosome position
# 4 SNP_7          1       19
# 5 SNP_8          1       23
# 2 SNP_5          1       34
# 3 SNP_6          1       41
# 1 SNP_4          1       60
#
# $SNPs$`2`
# ID chromosome position
# 3 SNP_11          2        2
# 2 SNP_10          2       43
# 1  SNP_9          2       63
#
# $SNPs$X
# ID chromosome position
# 2 SNP_13          X        6
# 6 SNP_17          X       15
# 4 SNP_15          X       39
# 1 SNP_12          X       79
# 3 SNP_14          X       95
# 5 SNP_16          X       97
#
# $SNPs$Y
# ID chromosome position
# 1 SNP_18          Y       32
# 2 SNP_19          Y       56
# 3 SNP_20          Y       88
#
#
# attr(,"class")
# [1] "geno.map"
DrawAFrame(map.sorted, gap = 30, ylim = c(1,3))
#   chr.id chr.is.allosome chr.length chr.start chr.end chr.mid
# 1      1           FALSE         60         1      61    31.0
# 2      2           FALSE         63        91     154   122.5
# X      X            TRUE         97       184     281   232.5
# Y      Y            TRUE         88       311     399   355.0
}

