% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOOLS_FIMPUTE.R
\name{Fimpute_geno_file}
\alias{Fimpute_geno_file}
\title{generate FIMPUTE genotype input file}
\usage{
Fimpute_geno_file(genodata, coding_AA = "AA", coding_AB = "AB",
  coding_BB = "BB", na_input = NULL, chip = NULL, PATH = "",
  title = "NewImp")
}
\arguments{
\item{genodata}{the genotype data as a matrix}

\item{coding_AA}{the code for genotype AA in the input file}

\item{coding_AB}{the code for genotype AB in the input file}

\item{coding_BB}{the code for genotype BB in the input file}

\item{na_input}{the code for genotype missing in the input file}

\item{chip}{a vector of integer for the chip value}

\item{PATH}{a path for the output}

\item{title}{a character vallue for the file name}
}
\value{
the data to be written into the data file.
}
\description{
DESCRIPTION
}
\note{
a wraper to generate FIMPUTE input genotype file
}
