% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIMPUTE.R
\name{Fimpute_map_file}
\alias{Fimpute_map_file}
\title{generate FIMPUTE map input file}
\usage{
Fimpute_map_file(geno_map, chip = NULL, PATH = "", title = "NewImp")
}
\arguments{
\item{geno_map}{a geno.map object}

\item{chip}{not support yet}

\item{PATH}{a path for the output}

\item{title}{a character vallue for the file name}
}
\value{
the data to be written into the data file.
}
\description{
DESCRIPTION
}
\note{
a wraper to generate FIMPUTE input map file
}
