% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOOLS_FIMPUTE.R
\name{Fimpute_read_geno}
\alias{Fimpute_read_geno}
\title{Read genotype file generated by FImpute}
\usage{
Fimpute_read_geno(imputed_file, map_file, ex_snp_file = NULL)
}
\arguments{
\item{imputed_file}{the genotype file generated by FImpute}

\item{map_file}{the map file used by FImpute}

\item{ex_snp_file}{the ex_snp file generated by FImpute?}
}
\value{
a matrix of genotypes (0/1/2 output)
}
\description{
DESCRIPTION
}
\note{
Read genotype file generated by FImpute
}
