% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{geno.AB2Num}
\alias{geno.AB2Num}
\alias{geno.Num2AB}
\title{transform the coding between BB/AB/AA and -1/0/1}
\usage{
geno.AB2Num(genodata, na.geno = NULL)

geno.Num2AB(genodata.int, code = c(1, 0, -1), na.geno.int = NULL)
}
\arguments{
\item{genodata}{A matrix of bi-allelic genotype (n(nind) x m(nmar))}

\item{na.geno}{A 2-digit character for missing genotype in monoAlle}

\item{genodata.int}{A matrix of integer for genotype}

\item{code}{A vector of integer for AA/AB/BB}

\item{na.geno.int}{A integer for missing genotype}
}
\value{
A matrix of integer

A matrix of bi-allelic genotype
}
\description{
transform the coding between BB/AB/AA and -1/0/1
}
\note{
This function transforms the coding from AA/AB/BB to -1/0/1

This function transforms the coding from -1/0/1 to BB/AB/AA
}
\examples{
genodata = rbind(c("AA","AB","--","CC")
                ,c("AA","--","BB","TT")
                ,c("AB","AB","AB","--"))
monoAlle = geno.biAL2monoAL(genodata)
monoAlle.AB = geno.AGCT2AB(monoAlle)
geno.AB = geno.monoAL2biAL(monoAlle.AB)
(mat.AB2Num = geno.AB2Num(geno.AB,"--"))
#    C1 C2 C3 C4
# R1  1  0 NA  1
# R2  1 NA  1 -1
# R3  0  0  0 NA
genodata = rbind(c("AA","AB","--","CC")
                ,c("AA","--","BB","TT")
                ,c("AB","AB","AB","--"))
monoAlle = geno.biAL2monoAL(genodata)
monoAlle.AB = geno.AGCT2AB(monoAlle)
geno.AB = geno.monoAL2biAL(monoAlle.AB)
(mat.AB2Num = geno.AB2Num(geno.AB,"--"))
#    C1 C2 C3 C4
# R1  1  0 NA  1
# R2  1 NA  1 -1
# R3  0  0  0 NA

geno.Num2AB(geno.AB2Num(geno.AB,"--"))
# C1   C2   C3   C4
# R1 "AA" "AB" NA   "AA"
# R2 "AA" NA   "AA" "BB"
# R3 "AB" "AB" "AB" NA

a = geno.AB2Num(geno.AB,"--")
a[is.na(a)] = -9
geno.Num2AB(a,na.geno = -9)
# C1   C2   C3   C4
# R1 "AA" "AB" NA   "AA"
# R2 "AA" NA   "AA" "BB"
# R3 "AB" "AB" "AB" NA
}

