% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{geno.AGCT2AB}
\alias{geno.AGCT2AB}
\title{Transform the coding of monoAL data to (A/B/NA).output}
\usage{
geno.AGCT2AB(monoAlle, na.output = 0)
}
\arguments{
\item{monoAlle}{A monoAL object}

\item{na.output}{A integer for out put data}
}
\value{
A monoAL_AB object
}
\note{
This function transforms the coding of monoAL data to (A/B/NA) 
             output, which is the index in the allele list / missing.
}
\examples{
   
genodata = rbind(c("AA","AB","--","CC")
                ,c("AA","--","BB","TT")
                ,c("AB","AB","AB","--"))
monoAlle = geno.biAL2monoAL(genodata)
(monoAlle.AB = geno.AGCT2AB(monoAlle))
# $dim
# Row Col 
# 3   4 
# 
# $names.ind
# [1] "R1" "R2" "R3"
# 
# $names.mar
# [1] "C1" "C2" "C3" "C4"
# 
# $names.base
# $names.base$C1
# [1] "A" "B"
# 
# $names.base$C2
# [1] "A" "B"
# 
# $names.base$C3
# [1] "B" "A"
# 
# $names.base$C4
# [1] "C" "T"
# 
# 
# $Allele_1
# [,1] [,2] [,3] [,4]
# [1,] "A"  "A"  "0"  "A" 
# [2,] "A"  "0"  "A"  "B" 
# [3,] "A"  "A"  "B"  "0" 
# 
# $Allele_2
# [,1] [,2] [,3] [,4]
# [1,] "A"  "B"  "0"  "A" 
# [2,] "A"  "0"  "A"  "B" 
# [3,] "B"  "B"  "A"  "0" 
# 
# attr(,"class")
# [1] "monoAL"    "monoAL_AB"
}
