% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{geno.biAL2monoAL}
\alias{geno.biAL2monoAL}
\title{Divide bi-allelic genotype data into mono-allelic}
\usage{
geno.biAL2monoAL(genodata, na.geno = "--", na.output = "00")
}
\arguments{
\item{genodata}{A matrix of bi-allelic genotype (n(nind) x m(nmar))}

\item{na.geno}{A 2-digit character for missing genotype in genedata}

\item{na.output}{A 2-digit character for out put data}
}
\value{
A monoAlle onject
}
\note{
This function divides bi-allelic genotype matrix into a
             mono-allelic object
}
\examples{
   
genodata = rbind(c("AA","AB","--","CC")
                ,c("AA","--","BB","TT")
                ,c("AB","AB","AB","--"))
(monoAlle = geno.biAL2monoAL(genodata))
#
# $dim
# Row Col 
# 3   4 
# 
# $names.ind
# [1] "R1" "R2" "R3"
# 
# $names.mar
# [1] "C1" "C2" "C3" "C4"
# 
# $names.base
# $names.base$C1
# [1] "A" "B"
# 
# $names.base$C2
# [1] "A" "B"
# 
# $names.base$C3
# [1] "B" "A"
# 
# $names.base$C4
# [1] "C" "T"
# 
# 
# $Allele_1
# C1  C2  C3  C4 
# R1 "A" "A" "0" "C"
# R2 "A" "0" "B" "T"
# R3 "A" "A" "A" "0"
# 
# $Allele_2
# C1  C2  C3  C4 
# R1 "A" "B" "0" "C"
# R2 "A" "0" "B" "T"
# R3 "B" "B" "B" "0"
# 
# attr(,"class")
# [1] "monoAL"
# Warning messages:
# 1: In geno.biAL2monoAL(genodata) :
#   no valid rownames for the input data, simple names will be given.
# 2: In geno.biAL2monoAL(genodata) :
#   no valid colnames for the input data, simple names will be given.
}

