% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEN_GenomicData.R
\name{geno.impu.ave}
\alias{geno.impu.ave}
\title{Impute the missing cells with average}
\usage{
geno.impu.ave(genodata)
}
\arguments{
\item{genodata}{A numeric matrix of genotype (missing coded as NA)}
}
\value{
A numeric matrix of genotype
}
\description{
Impute the missing cells with average
}
\note{
This function imputes missing value in genotype data using
             the average value of that column. This method can only be
             used for genotype data coded as integer (-1/0/1 or 0/1/2 or
             -10/0/10). It is usually used in regression methods
}
\examples{
genodata = rbind(c("AA","AB","--","CC")
                ,c("AA","--","BB","TT")
                ,c("AB","AB","AB","--"))
monoAlle = geno.biAL2monoAL(genodata)
monoAlle.AB = geno.AGCT2AB(monoAlle)
geno.AB = geno.monoAL2biAL(monoAlle.AB)
(mat.AB2Num = geno.AB2Num(geno.AB,"--"))
# C1 C2 C3 C4
# R1  1  0 NA  1
# R2  1 NA  1 -1
# R3  0  0  0 NA
geno.impu.ave(mat.AB2Num)
#    C1 C2  C3 C4
# R1  1  0 0.5  1
# R2  1  0 1.0 -1
# R3  0  0 0.0  0
}

