% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{geno.mono2PED}
\alias{geno.mono2PED}
\title{Create PED files}
\usage{
geno.mono2PED(monodata, IndID, lite = T, FamilyID = NULL, PID = NULL,
  MID = NULL, Sex = NULL, Pheno = NULL)
}
\arguments{
\item{monodata}{A monoAL object}

\item{IndID}{A vector of Individual ID}

\item{lite}{If using a lite version of PED file}

\item{FamilyID}{A vector of family ID}

\item{PID}{A vector of paternal ID}

\item{MID}{A vector of maternal ID}

\item{Sex}{A vector of individual sex}

\item{Pheno}{A vector of phenotypic value}
}
\value{
A data frame that can be written as a PED file
}
\note{
This function does some simple tests and combine available data
}
