% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEN_GenomicData.R
\name{geno.monoAL2biAL}
\alias{geno.monoAL2biAL}
\title{combine the allele data into Bi-allele data}
\usage{
geno.monoAL2biAL(monoAlle, na.geno = "00", na.output = "--")
}
\arguments{
\item{monoAlle}{A monoAL object}

\item{na.geno}{A 2-digit character for missing genotype in monoAlle}

\item{na.output}{A 2-digit character for out put data}
}
\value{
A matrix of bi-allelic genotype
}
\description{
combine the allele data into Bi-allele data
}
\note{
This function combines the allele data into Bi-allele data
}
\examples{
genodata = rbind(c("AA","AB","--","CC")
                ,c("AA","--","BB","TT")
                ,c("AB","AB","AB","--"))
monoAlle = geno.biAL2monoAL(genodata)
geno.monoAL2biAL(monoAlle)
#    C1   C2   C3   C4
# R1 "AA" "AB" "--" "CC"
# R2 "AA" "--" "BB" "TT"
# R3 "AB" "AB" "AB" "--"
}

