% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEN_GenomicData.R
\name{geno.qc}
\alias{geno.qc}
\title{Quality control for genotype data}
\usage{
geno.qc(genodata, code = c(1, 0, -1), na.string = NULL)
}
\arguments{
\item{genodata}{A matrix of genotype}

\item{code}{A vector of codes used in the genotype matrix}

\item{na.string}{The character used to code the missing value.}
}
\value{
A data.frame including basic statistics for SNPs.
}
\description{
Quality control for genotype data
}
\note{
Quality control for genotype data
}
\examples{
genodata_2 = rbind(c("AA","AB","--","AA","--")
                  ,c("AA","--","BB","BB","--")
                  ,c("AB","AB","AB","--","--"))

geno.qc(genodata_2,c("AA","AB","BB"),"--")
#   ngeno nmiss     rmiss NumAA NumAB NumBB       AAF       BAF       MAF
# 1     3     0 0.0000000     2     1     0 0.8333333 0.1666667 0.1666667
# 2     2     1 0.3333333     0     2     0 0.5000000 0.5000000 0.5000000
# 3     2     1 0.3333333     0     1     1 0.2500000 0.7500000 0.2500000
# 4     2     1 0.3333333     1     0     1 0.5000000 0.5000000 0.5000000
# 5     0     3 1.0000000     0     0     0 0.0000000 0.0000000 0.0000000
}

