% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_genomic_map.R
\name{geno_map_check_dup}
\alias{geno_map_check_dup}
\title{check duplicated positions}
\usage{
geno_map_check_dup(geno_map)
}
\arguments{
\item{geno_map}{a geno_map object}
}
\value{
a geno_map object
}
\description{
DESCRIPTION
}
\note{
check and randomly remove duplicated positions.
}
