% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenoR.R
\docType{package}
\name{GenoR}
\alias{GenoR}
\alias{GenoR-package}
\title{Management of genomic data}
\description{
GenoR is a package of functions for genomic data management
}
\author{
Tianfu Yang \email{tianfu@ualberta.ca}

Maintainer: Tianfu Yang \email{tianfu@ualberta.ca}
}

