% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.lsmeans}
\alias{plot.lsmeans}
\title{plot (multiple) lsmeans results}
\usage{
\method{plot}{lsmeans}(res.lm, SNPnames, nSNPs = length(SNPnames),
  mfrow = c(1, nSNPs), mar = c(4.1, 5.1, 2.1, 1.1), ylim = c(2, 6),
  ylab = "lsmean", xlab = NULL)
}
\arguments{
\item{res.lm}{result of lm()}

\item{SNPnames}{ID of SNPs}

\item{nSNPs}{number of SNPs}

\item{mfrow}{plot parameter. See par()}

\item{mar}{plot parameter. See par()}

\item{ylim}{The y-axis range}

\item{xlab,ylab}{Parameters passed to plot.default()}
}
\description{
plot (multiple) lsmeans results
}
\note{
This is a function to plot (multiple) lsmeans results.
#
}
\examples{
\\
}

