% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLINK.R
\name{read_plink_model_fisher}
\alias{read_plink_model_fisher}
\alias{read_plink_linear}
\title{read the result of plink association}
\usage{
read_plink_model_fisher(file, map, png_prefix = "./plink_model_fisher_manh_")

read_plink_linear(file, map, png_prefix = "./plink_linear_manh_")
}
\arguments{
\item{file}{the output file from plink}

\item{map}{the map of SNPs}

\item{png_prefix}{the prefix of the output Manhattan plot}

\item{file}{the output file from plink}

\item{map}{the map of SNPs}

\item{png_prefix}{the prefix of the output Manhattan plot}
}
\value{
a list of the association result

a list of the association result
}
\description{
DESCRIPTION

DESCRIPTION
}
\note{
read the result of plink association - calculate Q value and 
             generate Manhattan plot

read the result of plink association - calculate Q value and 
             generate Manhattan plot
}
