% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repmat}
\alias{repmat}
\title{Replication of matrix (mat) n x m times}
\usage{
repmat(mat, nrow, ncol)
}
\arguments{
\item{mat}{A matrix}

\item{nrow}{Number of rows}

\item{ncol}{Number of cols}
}
\value{
A matrix
}
\note{
This function is an application of kronecker multiplication.
}
\examples{
   
mat1 = matrix(1:4,2,2)
repmat(mat1,2,3)
#      [,1] [,2] [,3] [,4] [,5] [,6]
# [1,]    1    3    1    3    1    3
# [2,]    2    4    2    4    2    4
# [3,]    1    3    1    3    1    3
# [4,]    2    4    2    4    2    4
}

