% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ScaleMatrics}
\alias{ScaleMatrics}
\title{Matrics scaling}
\usage{
ScaleMatrics(mat, scale.coef, MARGIN)
}
\arguments{
\item{mat}{A numeric matrix (m * n)}

\item{scale.coef}{A numeric factor (m or n)}

\item{MARGIN}{An integer:
1 for row manipulation and 2 for column manipulation}
}
\value{
An sparse numeric matrix m * n
}
\note{
Each vector in the matrix is scaled using the corresponding
             element in the scale.coef.
             Left multiplication manipulate the rows, while right 
             multiplication manipulate the cols. This function uses diagonal 
             sparse functions to multiply each row/column by a nonzero 
             scalar (can be different across rows/columns)
             Thanks for Dr. Zhiquan Wang's contribution.
}
\examples{
(mat = matrix(1,3,5))
[,1] [,2] [,3] [,4] [,5]
[1,]    1    1    1    1    1
[2,]    1    1    1    1    1
[3,]    1    1    1    1    1

ScaleMatrics(mat,c(2,3,5),1)
3 x 5 Matrix of class "dgeMatrix"
     [,1] [,2] [,3] [,4] [,5]
[1,]    2    2    2    2    2
[2,]    3    3    3    3    3
[3,]    5    5    5    5    5

ScaleMatrics(mat,c(2,3,5,7,8),2)
# 3 x 3 Matrix of class "dgeMatrix"
# [,1] [,2] [,3]
# [1,]    2    3    5
# [2,]    2    3    5
# [3,]    2    3    5
}

