% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_map.R
\name{SNP2Window}
\alias{SNP2Window}
\title{# SNP2Window(SID, chromosome, position, distance = 30)
# $Chromosomes
# chr.id chr.is.allosome chr.length chr.n.window
# 0      0           FALSE         77            3
# 1      1           FALSE         98            4
# 2      2           FALSE         88            3
# X      X            TRUE         94            4
# Y      Y            TRUE         68            3
# 
# $SNPs
# $SNPs$`0`
# ID chromosome position window
# 2 SNP_2          0       41    0_2
# 3 SNP_3          0       72    0_3
# 1 SNP_1          0       77    0_3
# 
# $SNPs$`1`
# ID chromosome position window
# 3 SNP_6          1       18    1_1
# 4 SNP_7          1       22    1_1
# 5 SNP_8          1       23    1_1
# 1 SNP_4          1       51    1_2
# 2 SNP_5          1       98    1_4
# 
# $SNPs$`2`
# ID chromosome position window
# 1  SNP_9          2       27    2_1
# 2 SNP_10          2       52    2_2
# 3 SNP_11          2       88    2_3
# 
# $SNPs$X
# ID chromosome position window
# 6 SNP_17          X       11    X_1
# 5 SNP_16          X       34    X_2
# 3 SNP_14          X       37    X_2
# 2 SNP_13          X       60    X_2
# 1 SNP_12          X       65    X_3
# 4 SNP_15          X       94    X_4
# 
# $SNPs$Y
# ID chromosome position window
# 1 SNP_18          Y        2    Y_1
# 3 SNP_20          Y       44    Y_2
# 2 SNP_19          Y       68    Y_3
# 
# attr(,"class")
# [1] "geno.snp2window" "geno.map"}
\usage{
SNP2Window(SID, chromosome, position, allosome = NULL, distance = 10^6)
}
\description{
# SNP2Window(SID, chromosome, position, distance = 30)
# $Chromosomes
# chr.id chr.is.allosome chr.length chr.n.window
# 0      0           FALSE         77            3
# 1      1           FALSE         98            4
# 2      2           FALSE         88            3
# X      X            TRUE         94            4
# Y      Y            TRUE         68            3
# 
# $SNPs
# $SNPs$`0`
# ID chromosome position window
# 2 SNP_2          0       41    0_2
# 3 SNP_3          0       72    0_3
# 1 SNP_1          0       77    0_3
# 
# $SNPs$`1`
# ID chromosome position window
# 3 SNP_6          1       18    1_1
# 4 SNP_7          1       22    1_1
# 5 SNP_8          1       23    1_1
# 1 SNP_4          1       51    1_2
# 2 SNP_5          1       98    1_4
# 
# $SNPs$`2`
# ID chromosome position window
# 1  SNP_9          2       27    2_1
# 2 SNP_10          2       52    2_2
# 3 SNP_11          2       88    2_3
# 
# $SNPs$X
# ID chromosome position window
# 6 SNP_17          X       11    X_1
# 5 SNP_16          X       34    X_2
# 3 SNP_14          X       37    X_2
# 2 SNP_13          X       60    X_2
# 1 SNP_12          X       65    X_3
# 4 SNP_15          X       94    X_4
# 
# $SNPs$Y
# ID chromosome position window
# 1 SNP_18          Y        2    Y_1
# 3 SNP_20          Y       44    Y_2
# 2 SNP_19          Y       68    Y_3
# 
# attr(,"class")
# [1] "geno.snp2window" "geno.map"
}

