% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genomic_map.R
\name{SortMap}
\alias{SortMap}
\title{build a geno_map object with sorted information}
\usage{
SortMap(SID, chromosome, position, allosome = NULL)
}
\arguments{
\item{SID}{A vector of SNP IDs (character)}

\item{chromosome}{A vector of chromosome (character) where the SNP
is located. Support [1-9]{1,2}|[A-Z] ie. 0-99 or
a single upper letter. 0 for unmapped SNPs}

\item{position}{A vector of position (integer) where the SNP
is located.}

\item{allosome}{A vector of the names of allosomes. By default all
chromosomes natated with upper letters will be
treated as allosomes}
}
\value{
A geno.map object
}
\description{
build a geno_map object with sorted information
}
\note{
This is a function for sortting map information of SNPs.
}
\examples{
chromname = c(0:2,"X","Y")
SID = paste("SNP",1:20,sep="_")
chromosome = c(rep("0",3),rep("1",5),rep("2",3),rep("X",6),rep("Y",3)))
position = sample(1:100,20)

data.frame(SID,chromosome,position)
#       SID chromosome position
# 1   SNP_1          0       31
# 2   SNP_2          0       71
# 3   SNP_3          0       45
# 4   SNP_4          1       60
# 5   SNP_5          1       34
# 6   SNP_6          1       41
# 7   SNP_7          1       19
# 8   SNP_8          1       23
# 9   SNP_9          2       63
# 10 SNP_10          2       43
# 11 SNP_11          2        2
# 12 SNP_12          X       79
# 13 SNP_13          X        6
# 14 SNP_14          X       95
# 15 SNP_15          X       39
# 16 SNP_16          X       97
# 17 SNP_17          X       15
# 18 SNP_18          Y       32
# 19 SNP_19          Y       56
# 20 SNP_20          Y       88

SortMap(SID, chromosome, position)
# $Chromosomes
#   chr.id chr.is.allosome chr.length
# 0      0           FALSE         71
# 1      1           FALSE         60
# 2      2           FALSE         63
# X      X            TRUE         97
# Y      Y            TRUE         88
#
# $SNPs
# $SNPs$`0`
#    ID    chromosome position
# 1 SNP_1           0       31
# 3 SNP_3           0       45
# 2 SNP_2           0       71
#
# $SNPs$`1`
#    ID    chromosome position
# 4 SNP_7           1       19
# 5 SNP_8           1       23
# 2 SNP_5           1       34
# 3 SNP_6           1       41
# 1 SNP_4           1       60
#
# $SNPs$`2`
#    ID    chromosome position
# 3 SNP_11          2        2
# 2 SNP_10          2       43
# 1  SNP_9          2       63
#
# $SNPs$X
#    ID    chromosome position
# 2 SNP_13          X        6
# 6 SNP_17          X       15
# 4 SNP_15          X       39
# 1 SNP_12          X       79
# 3 SNP_14          X       95
# 5 SNP_16          X       97
#
# $SNPs$Y
#    ID    chromosome position
# 1 SNP_18          Y       32
# 2 SNP_19          Y       56
# 3 SNP_20          Y       88
#
#
# attr(,"class")
# [1] "geno.map"
}

