% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{Stat.qvalue}
\alias{Stat.qvalue}
\title{calculate q-values based on a set of p-values in Multiple Tests}
\usage{
Stat.qvalue(p.value)
}
\arguments{
\item{p.value}{The p-values obtained from multiple testing}

\item{\}{}

\item{\}{}

\item{\}{}

\item{\}{}

\item{\}{}
}
\value{
\item{name a}{description} \
}
\description{
calculate q-values based on a set of p-values in Multiple Tests
}
\note{
This is a revised version of the qvalue() function in
             R/rrBLUP::GWAS().
}
\examples{
p = runif(1000)
q2 = Stat.qvalue(p)
}
\references{
Estimating the False Discovery Rate using SAS (Paper 190-31)
}

