% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OTHER_utils.R
\name{strsplit.mat}
\alias{strsplit.mat}
\title{A matrix version of strsplit()}
\usage{
strsplit.mat(vec.char, sep = " ")
}
\arguments{
\item{vec.char}{A vector of characters which will be split.}

\item{sep}{The separator passed to strsplit()}
}
\value{
A matrix
}
\description{
A matrix version of strsplit()
}
\note{
Instead of outputing a list, this func try to return a matrix.
}
\examples{
vec.char = c("QWERTYUIOP",
             "ASDFGHJKL:",
             "ZXCVBNM<>?")
strsplit.mat(vec.char, "")

#      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
# [1,] "Q"  "W"  "E"  "R"  "T"  "Y"  "U"  "I"  "O"  "P"
# [2,] "A"  "S"  "D"  "F"  "G"  "H"  "J"  "K"  "L"  ":"
# [3,] "Z"  "X"  "C"  "V"  "B"  "N"  "M"  "<"  ">"  "?"
}

