% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOOLS_PLINK.R
\name{write_ped_file}
\alias{write_ped_file}
\title{write a ped file}
\usage{
write_ped_file(ped_file, file)
}
\arguments{
\item{ped_file}{a ped_file object generated by geno.mono2PED}

\item{file}{path of the output file}
}
\description{
DESCRIPTION
}
\note{
Write a ped file based on a ped_file object generated by geno.mono2PED
}
