% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OTHER_utils.R
\name{clc.functions}
\alias{clc.functions}
\title{clean all (customed) functions in present environment}
\usage{
clc.functions()
}
\description{
clean all (customed) functions in present environment
}
\note{
This function looks for all functions defined in current
             session and cleaning them. It may be useful when a
             environment image is loaded and functions may be masked.
}

