% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{geno.monoAL2Num}
\alias{geno.monoAL2Num}
\title{transform the coding from mono-allele to integer (-1/0/1/-5)}
\usage{
geno.monoAL2Num(monoAlle_AB, na.geno = 0, na.output = -5)
}
\arguments{
\item{monoAlle_AB}{A monoAL_AB object}

\item{na.geno}{An integer for missing genotype in monoAlle}

\item{na.output}{An integer for out put data}
}
\value{
A matrix of integer
}
\note{
This function transforms the coding from mono-allele to 
             integer (-1/0/1/-5)
}
\examples{
   
genodata = rbind(c("AA","AB","--","CC")
                ,c("AA","--","BB","TT")
                ,c("AB","AB","AB","--"))
monoAlle = geno.biAL2monoAL(genodata)
monoAlle.AB = geno.AGCT2AB(monoAlle)
(mat.int = geno.monoAL2Num(monoAlle.AB))
#    C1 C2 C3 C4
# R1  1  0 -5  1
# R2  1 -5  1 -1
# R3  0  0  0 -5
}
