% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Manhattan}
\alias{Manhattan}
\title{Draw Manhattan plots}
\usage{
Manhattan(res.gwas.ID, res.gwas.value, map.sorted, gap = 5e+07, ylim = NULL,
  xlab = "Chromosome", ylab = "", title = "", axes = TRUE,
  ltype = "p", pch = 20, cols = c("dark blue", "cornflowerblue"))
}
\arguments{
\item{res.gwas.ID}{The ID of markers from GWAS result}

\item{res.gwas.value}{The value of markers from GWAS result}

\item{map.sorted}{A sorted map}

\item{gap}{The gap between chromosomes in the plot}

\item{ylim}{The y-axis range}

\item{xlab, ylab}{Parameters passed to plot.default()}

\item{title}{Character, the title of the plot}

\item{axes}{Logical, draw the axes or not}

\item{ltype}{"lty" passed to plot.default()}

\item{pch}{Type of points}

\item{cols}{A Vector with 2 elements of color. can be used
to distinguish two adjacent chromosomes.}
}
\value{
A data frame of chromosome information (from DrawAFrame())
}
\note{
This is a function that draws Manhattan plots
}

