% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OTHER_utils.R
\name{ScaleMatrics}
\alias{ScaleMatrics}
\title{Matrics scaling}
\usage{
ScaleMatrics(mat, scale.coef, MARGIN)
}
\arguments{
\item{mat}{A numeric matrix (m * n)}

\item{scale.coef}{A numeric factor (m or n)}

\item{MARGIN}{An integer:
1 for row manipulation and 2 for column manipulation}
}
\value{
An sparse numeric matrix m * n
}
\description{
Matrics scaling
}
\note{
Each vector in the matrix is scaled using the corresponding
             element in the scale.coef.
             Left multiplication manipulate the rows, while right
             multiplication manipulate the cols. This function uses diagonal
             sparse functions to multiply each row/column by a nonzero
             scalar (can be different across rows/columns)
             Thanks for Dr. Zhiquan Wang's contribution.
}
\examples{
mat = matrix(1,3,3)
ScaleMatrics(mat,c(2,3,5),1)
# 3 x 3 Matrix of class "dgeMatrix"
# [,1] [,2] [,3]
# [1,]    2    2    2
# [2,]    3    3    3
# [3,]    5    5    5

ScaleMatrics(mat,c(2,3,5),2)
# 3 x 3 Matrix of class "dgeMatrix"
# [,1] [,2] [,3]
# [1,]    2    3    5
# [2,]    2    3    5
# [3,]    2    3    5
}

