% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Stat.qvalue}
\alias{Stat.qvalue}
\title{calculate q-values based on a set of p-values in Multiple Tests}
\usage{
Stat.qvalue(p.value)
}
\arguments{
\item{p.value}{The p-values obtained from multiple testing}
}
\value{
q.value with the same length of p.value
}
\note{
This is a revised version of the qvalue() function in
             R/rrBLUP::GWAS().
}
\examples{
p = runif(1000)
q2 = Stat.qvalue(p)
}
\references{
Estimating the False Discovery Rate using SAS (Paper 190-31)
}

